+++ pyproject.toml
@@ -13,11 +13,15 @@ dependencies = [
 sbutterfly = "main:main"
 
 [tool.setuptools]
-packages = [
-    "interfaces",
-    "plugins",
-]
-py-modules = ["main"]
+package-dir = { "" = "src" }
+py-modules = ["main", "utils"]
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
 
 
 [tool.uv]
